Integrated File.requestPermission()
Integrated the API "Integrated File.requestPermission()" in hope that it might catch denied access situations on macOS (but to no avail).

However, the code seems to be working well (not worse than before the change at least) so I will let it be.

Also fixed a bug: accessing system folders (such as /etc) previously crashed the file browser. Now it recovers gracefully, showing a message that access to that folder is denied.